impr: Remove `/poll` (#642)
Also refactors the code to be focused on QOTDs

The main downside of switching to Discord polls (for us) is that they have an expiry, Scradd's don't and were able to run indefinitely. If this is too much of a trade-off, I can keep `/poll`. Let me know what you think.

Old polls with `vote-mode: True` will still be enforced.

closes #634 
closes #333